Makefile: allow $(INSTALL) to be overriden
Teach the Makefile to follow the usual convention of using
"$(INSTALL) -d" to create directories, and "$(INSTALL) [-m mode]
files... dir" to install files, instead of hardcoding program names
for "mkdir" and "install".

With this, enhanced "instsall" program can be more easily used with
the build system. For example, "nstalli" that has "--dry-run" option
to tell what would happen can be used like this:

  $ make INSTALL="nstalli --dry-run" install

Also, "install" target should depend on "all" which builds cvsps
itself, which is one of the things it tries to install.  Make it so.

Signed-off-by: Junio C Hamano <gitster@pobox.com>